% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudflareapi-package.R
\docType{package}
\name{cloudflareapi-package}
\alias{cloudflareapi}
\alias{cloudflareapi-package}
\title{cloudflareapi: 'Cloudflare' 'API' Client}
\description{
A client to interact with the 'Cloudflare' 'API' described at \url{https://developers.cloudflare.com/api/}. The 'Cloudflare' 'API' provides programmatic access to 'Cloudflare.com', a platform that offers content delivery, security, and performance optimization services for websites and applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/api2r/cloudflareapi}
  \item \url{https://api2r.github.io/cloudflareapi/}
  \item Report bugs at \url{https://github.com/api2r/cloudflareapi/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
